commit 3

+++ src/app/contracts/contract-list/contract-list.component.ts
@@ -1,31 +1,70 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
+import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import jsPDF from 'jspdf';
 import { Contrat } from '../../core/Model/Contrat';
 import { ContratService } from '../../core/services/contrat.service';
+import { Observable, VirtualTimeScheduler } from 'rxjs';
+import {
+  ApexChart,
+  ApexDataLabels,
+  ApexNonAxisChartSeries,
+  ApexTitleSubtitle,
+} from 'ng-apexcharts';
 
+import {
+  map,
+  filter,
+  debounceTime,
+  distinctUntilChanged,
+  tap,
+  switchMap,
+} from 'rxjs/operators';
+import { EtudiantService } from 'src/app/core/services/etudiant.service';
+import { Etudiant } from 'src/app/core/Model/Etudiant';
+import { NotifierService } from 'angular-notifier';
 @Component({
   selector: 'app-contract-list',
   templateUrl: './contract-list.component.html',
   styleUrls: ['./contract-list.component.css'],
 })
 export class ContractListComponent implements OnInit {
+  searchText: any;
+  parentData =""
   title = 'pagination';
   page: number = 1;
   count: number = 0;
-  tableSize: number = 1;
-  tableSizes: any = [1, 2, 15, 20];
+  tableSize: number = 5;
+  tableSizes: any = [1, 5, 10, 15];
+  order: boolean = true;
+  queryField = new FormControl();
+  results$: Observable<any>;
+  total = 0;
+	private notifier: NotifierService;
 
   @ViewChild('content', { static: false }) el!: ElementRef;
 
-  constructor(private contratService: ContratService) {}
+  constructor(
+    private contratService: ContratService,
+    private fb: FormBuilder,
+    notifier: NotifierService
+  ) {this.notifier= notifier}
   contrat: Contrat;
-
+  searchForm: FormGroup;
   contratList: Contrat[];
+  nbr: any;
+  m = new Date().getMonth() + 1;
+  d = new Date();
   ngOnInit(): void {
+    console.log(this.m);
+
     this.contrat = new Contrat();
     //getContracts
     this.contratService.getAllContrat().subscribe((data: Contrat[]) => {
       this.contratList = data;
+    });
+
+    this.searchForm = this.fb.group({
+      search: new FormControl(),
     });
   }
 
@@ -39,18 +78,18 @@ export class ContractListComponent implements OnInit {
   }
   //deleteContract
   delete(id: number) {
-    this.contratService.deleteContrat(id).subscribe((data) => {
-      console.log(data);
-      this.contratService.getAllContrat();
-      location.reload();
-    });
+    if (confirm('Are you sure to delete this Contract!'))
+      this.contratService.deleteContrat(id).subscribe((data) => {
+        alert('Contract deleted successfully!');
+        console.log(data);
+        this.contratService.getAllContrat();
+        location.reload();
+      });
   }
-  //updateContract
-  update() {}
 
   //generate pdf
   makePDF() {
-    let pdf = new jsPDF('p', 'pt', 'a4');
+    let pdf = new jsPDF('p', 'pt', 'a5');
     pdf.html(this.el.nativeElement, {
       callback: (pdf) => {
         pdf.save('contrat.pdf');
@@ -63,13 +102,61 @@ export class ContractListComponent implements OnInit {
     });
   }
   onTableDataChange(event: any): void {
-    this.page = event
+    this.page = event;
     this.postList();
   }
-  
-  onTableSizeChange(event: any){
+
+  onTableSizeChange(event: any) {
     this.tableSize = event.target.value;
     this.page = 1;
     this.postList();
   }
+
+  sort(event: any) {
+    switch (event.target.value) {
+      case 'Low': {
+        this.contratList = this.contratList.sort(
+          (low, high) => low.montantContrat - high.montantContrat
+        );
+        break;
+      }
+
+      case 'High': {
+        this.contratList = this.contratList.sort(
+          (low, high) => high.montantContrat - low.montantContrat
+        );
+        break;
+      }
+
+      default: {
+        this.contratList = this.contratList.sort(
+          (low, high) => low.montantContrat - high.montantContrat
+        );
+        break;
+      }
+    }
+    return this.contratList;
+  }
+
+  //ddlIsNear
+  ContractDateIsNear(id: number) {
+    return this.contratService.ContractDateIsNear(id);
+  }
+
+  transferData(){
+    this.parentData=this.searchText
+  }
+
+
+	/**
+	 * Show a notification
+	 *
+	 * @param {string} type    Notification type
+	 * @param {string} message Notification message
+	 */
+	public showNotification( type: string, message: string ): void {
+		this.notifier.notify( type, message );
+	}
 }
+
+
